refactor: extract settings key oauth and model routes
